refactor: [SHU-627] split query_service.py into mixin-based query/ package
Decompose the 1680-line monolithic QueryService into focused modules
organized by search type, improving maintainability and testability.

Changes:
- Extract SimilaritySearchMixin into query/similarity.py (200 lines)
- Extract KeywordSearchMixin into query/keyword.py (454 lines)
- Extract HybridSearchMixin into query/hybrid.py (241 lines)
- Extract MultiSurfaceSearchMixin into query/multi_surface.py (234 lines)
- Move shared utilities to QueryServiceBase in query/base.py (459 lines)
- Slim query_service.py to dispatcher + mixin composition (196 lines)
- Merge 008_0003 migration into 008_0001 (avoid revision conflict)
- Add idempotent SQL script for one-off column rename
